hotpatch: formatter-workbench update

+++ 3.0.2/formatter-workbench/index.js
@@ -1,19 +1,66 @@
 'use strict';
 
-var grid, tabBar, tutorial;
+var grid, tabBar, tutorial, callApi;
+
+var tutTOC = [
+    'Table of Contents.html',
+    'Welcome.html',
+    'The Workench Interface.html',
+    [
+        'The Hypergrid Section.html',
+        'The Editor Section.html',
+        'The Tutorial Section.html'
+    ],
+    'The Data Tab.html',
+    [
+        'Editing grid cells (Activity 1).html',
+        'Editing the data object (Activity 2).html',
+        'Add a new row to the data object (Activity 3).html'
+    ],
+    'The State Tab.html',
+    [
+        'Setting grid state (Activity 4).html',
+        'Setting column properties via grid state.html',
+        'Setting row and cell properties via grid state.html',
+        'The Properties Cascade.html',
+        'Edit a column property (Activity 5).html'
+    ],
+    'The Localizers Tab.html',
+    [
+        'Bind a cell to a localizer (Activity 6).html',
+        'Adding a new localizer.html',
+        'Create a Localizer (Activity 7).html'
+    ],
+    'The Cell Editors Tab (Activity 8).html',
+    [
+        'Activity 9.html'
+    ],
+    'Validation.html',
+    [
+        'The parse() method.html',
+        'The invalid() method.html',
+        'Activity 10 - Returning parsing errors.html'
+    ]
+];
+
+var svg = {
+    'svg-reset': '<svg viewBox="0 0 32 32" version="1.1" width="22" height="22"><path d="M 15.5 2.09375 L 14.09375 3.5 L 16.59375 6.03125 C 16.394531 6.019531 16.203125 6 16 6 C 10.5 6 6 10.5 6 16 C 6 17.5 6.304688 18.894531 6.90625 20.09375 L 8.40625 18.59375 C 8.207031 17.792969 8 16.898438 8 16 C 8 11.601563 11.601563 8 16 8 C 16.175781 8 16.359375 8.019531 16.53125 8.03125 L 14.09375 10.5 L 15.5 11.90625 L 19.71875 7.71875 L 20.40625 7 L 19.71875 6.28125 Z M 25.09375 11.90625 L 23.59375 13.40625 C 23.894531 14.207031 24 15.101563 24 16 C 24 20.398438 20.398438 24 16 24 C 15.824219 24 15.640625 23.980469 15.46875 23.96875 L 17.90625 21.5 L 16.5 20.09375 L 12.28125 24.28125 L 11.59375 25 L 12.28125 25.71875 L 16.5 29.90625 L 17.90625 28.5 L 15.40625 25.96875 C 15.601563 25.980469 15.804688 26 16 26 C 21.5 26 26 21.5 26 16 C 26 14.5 25.695313 13.105469 25.09375 11.90625 Z "></path></svg>',
+    'svg-delete': '<svg viewBox="0 0 12 16" version="1.1" width="12" height="16" class="dangerous-button"><path fill-rule="evenodd" d="M11 2H9c0-.55-.45-1-1-1H5c-.55 0-1 .45-1 1H2c-.55 0-1 .45-1 1v1c0 .55.45 1 1 1v9c0 .55.45 1 1 1h7c.55 0 1-.45 1-1V5c.55 0 1-.45 1-1V3c0-.55-.45-1-1-1zm-1 12H3V5h1v8h1V5h1v8h1V5h1v8h1V5h1v9zm1-10H2V3h9v1z"></path></svg>'
+};
 
 window.onload = function() {
     var NEW = '(New)';
+    var isCamelCase = /[a-z][A-Z]/;
     var saveFuncs = {
         editor: saveCellEditor,
         localizer: saveLocalizer
     };
     var defaults = {
         data: [
-            { symbol: 'APPL', name: 'Apple Inc.', prevclose: 93.13, change: .0725 },
+            { symbol: 'APPL', name: 'Apple Inc.', prevclose: 93.13, change: -.0725 },
             { symbol: 'MSFT', name: 'Microsoft Corporation', prevclose: 51.91, change: .0125 },
             { symbol: 'TSLA', name: 'Tesla Motors Inc.', prevclose: 196.40, change: .08 },
-            { symbol: 'IBM', name: 'International Business Machines Corp', prevclose: 155.35, change: .02375 }
+            { symbol: 'IBM', name: 'International Business Machines Corp', prevclose: 155.35, change: -.02375 }
         ],
         state: {
             showRowNumbers: false, // override the default (true)
@@ -25,34 +72,46 @@ window.onload = function() {
                 }
             }
         }
-    };
+    }
+
+    // Append version numbers to <h1> header
+    document.querySelector('body > h1:first-child').innerHTML += ' <sup>(rev. 11)</sup> — Hypergrid <sup>(v' + fin.Hypergrid.prototype.version + ')</sup>';
+
+    cloneSvgEls();
 
     grid = new fin.Hypergrid();
 
-    var tabBars = document.querySelectorAll('.curvy-tabs-container');
-
-    tabBar = new CurvyTabs(tabBars[0]);
+    tabBar = new CurvyTabs(document.getElementById('tab-bar-editors'));
     tabBar.paint();
 
-    tutorial = new Tutorial(tabBars[1], 'tutorial/', 19);
+    tutorial = new CurvyTabsPager(
+        document.getElementById('page-panel'),
+        new CurvyTabs(document.getElementById('tab-bar-tutorial')),
+        'tutorial/',
+        flatten(tutTOC)
+    );
 
-    initTextEditor('data');
-    initTextEditor('state');
+    callApi('data'); // inits both 'data' and 'state' editors
+
     initLocalsButtons();
 
-    Object.keys(scripts).forEach(function (key) {
-        initObjectEditor(key);
+    var scripts = {
+        editor: undefined,
+        localizer: undefined
+    };
+    Object.keys(scripts).forEach(function(type) {
+        ajax(type + 's.js', initObjectEditor.bind(null, type));
     });
 
-    document.getElementById('reset-all').onclick = function () {
-        if (confirm('Clear localStorage and reload?')) {
+    document.getElementById('reset-all').onclick = function() {
+        if (confirm('Clear localStorage and reload?\n\nThis will reset all tabs to their default values, removing all edits, including new custom localizers and custom cell editors.')) {
             localStorage.clear();
             location.reload();
         }
     };
 
-    grid.addEventListener('fin-after-cell-edit', function (e) {
-        putJSON('data', grid.behavior.getData());
+    grid.addEventListener('fin-after-cell-edit', function(e) {
+        document.getElementById('data').value = stringifyAndUnquoteKeys(grid.behavior.getData());
     });
 
     var dragger, divider = document.querySelector('.divider');
@@ -73,11 +132,12 @@ window.onload = function() {
                 newTabHeight = dragger.tabHeight - topDelta;
 
             if (newGridHeight >= 65 && newTabHeight >= 130) {
-                divider.style.borderTopStyle = divider.style.borderTopColor = null;
+                divider.style.borderTopStyle = divider.style.borderTopColor = null; // revert to :active style
                 divider.style.top = newDividerTop + 'px';
                 grid.div.style.height = (dragger.gridHeight = newGridHeight) + 'px';
                 tabBar.container.style.height = (dragger.tabHeight = newTabHeight) + 'px';
             } else {
+                // force :hover style when out of range even though dragging (i.e., :active)
                 divider.style.borderTopStyle = 'double';
                 divider.style.borderTopColor = '#444';
             }
@@ -90,33 +150,128 @@ window.onload = function() {
         dragger = undefined;
     });
 
-    function initTextEditor(type) {
-        document.getElementById(type).value = localStorage.getItem(type);
-        if (!document.getElementById(type).value) {
-            putJSON(type, defaults[type]);
-        }
-        var apiMethodName = 'set' + capitalize(type);
-        callApi(apiMethodName, type);
+    // Inject svg element into elements with svg-wildcard class name
+    function cloneSvgEls() {
+        querySelectorEach('[class*="svg-"]', function (el) {
+            el.innerHTML = svg[el.className.match(/\bsvg-[\w-]+/)[0]];
+        });
+    }
 
-        document.getElementById('reset-' + type).onclick = resetTextEditor;
+    function querySelectorEach(selector, iterator, context) {
+        var els = document.querySelectorAll(selector);
+        if (els.forEach) {
+            els.forEach(iterator, context);
+        } else {
+            Array.prototype.forEach.call(els, iterator, context);
+        }
+    }
+
+    function flatten(arr) {
+        var result = [];
+        walk(tutTOC);
+        function walk(list) {
+            list.forEach(function(item) {
+                if (Array.isArray(item)) {
+                    walk(item);
+                } else {
+                    result.push(item);
+                }
+            });
+        }
+        return result;
+    }
+
+    function ajax(url, callback) {
+        var httpRequest = new XMLHttpRequest();
+
+        httpRequest.open('GET', url);
+
+        httpRequest.onreadystatechange = function() {
+            if (
+                httpRequest.readyState === 4 && // HTTP_STATE_DONE
+                httpRequest.status === 200 // HTTP_STATUS_OK
+            ) {
+                callback(httpRequest.responseText);
+            }
+        };
+
+        httpRequest.send();
+    }
+
+    function callApi(methodName, type, confirmation) {
+        // When `methodName` is `undefined` or omitted promote 2nd and 3rd params
+        if (!methodName || !isCamelCase.test(methodName)) {
+            confirmation = type;
+            type = methodName;
+            methodName = 'set' + capitalize(type);
+        }
+
+        var texEl = document.getElementById(type); // tab editor's textarea element
+        if (!texEl.value && !(texEl.value = localStorage.getItem(type))) {
+            localStorage.setItem(type, texEl.value = stringifyAndUnquoteKeys(defaults[type]));
+        }
+        texEl.oninput = function() {
+            resetEl.classList.toggle('disabled', texEl.value === localStorage.getItem(type));
+        };
+
+        // We're using eval here instead of JSON.parse because we want to allow unquoted keys.
+        // Note: L-value must be inside eval because R-value beginning with '{' is eval'd as BEGIN block.
+        var value;
+        eval('value =' + texEl.value);
+
+        if (methodName === 'setData') {
+            grid.setData(value, { schema: [] });
+            callApi('state'); // reapply state after resetting schema (also inits state editor on first time called)
+        } else {
+            grid[methodName](value);
+        }
+
+        if (confirmation) {
+            feedback(texEl.parentElement, confirmation);
+        }
+
+        var resetEl = document.getElementById('reset-' + type);
+        texEl.oninput();
+        resetEl.onclick = resetTextEditor;
+
     }
 
     function resetTextEditor() {
         var type = this.id.replace(/^reset-/, '');
-        if (confirm('Reset the ' + capitalize(type) + ' tab editor to its default?')) {
+        if (!isDisabled(this) && confirm('Reset the ' + capitalize(type) + ' tab editor to its default?')) {
+            document.getElementById(type).value = '';
             localStorage.removeItem(type);
-            initTextEditor(type);
+            callApi(type);
         }
     }
 
-    function resetObjectEditor() {
+    function resetObject() {
         var type = this.id.replace(/^reset-/, '');
-        var key = document.getElementById(type + '-dropdown').value;
-        if (confirm('Reset the "' + key + '" ' + type + ' to its default?')) {
-            var script = scripts[type].find(isScriptByName.bind(null, key));
-            localStorage.setItem(type + '_' + key, script);
+        var name = document.getElementById(type + '-dropdown').value;
+        if (!isDisabled(this) && confirm('Reset the "' + name + '" ' + type + ' to its default?')) {
+            var script = getDefaultScript(type, name);
+            if (name !== NEW) {
+                localStorage.setItem(type + '_' + name, script);
+            }
             document.getElementById(type + '-script').value = script;
+            enableResetAndDeleteIcons(type, name);
         }
+    }
+
+    function deleteObject() {
+        var type = this.id.replace(/^delete-/, '');
+        var dropdown = document.getElementById(type + '-dropdown');
+        var name = dropdown.value;
+        if (!isDisabled(this) && confirm('Delete the "' + name + '" ' + type + '?')) {
+            dropdown.options.remove(dropdown.selectedIndex);
+            dropdown.selectedIndex = 0; // "(New)"
+            dropdown.onchange();
+            localStorage.removeItem(type + '_' + name);
+        }
+    }
+
+    function isDisabled(el) {
+        return el.classList.contains('disabled');
     }
 
     function capitalize(str) {
@@ -127,7 +282,7 @@ window.onload = function() {
         var el = document.getElementById(type + '-dropdown').parentElement.querySelector('.locals');
         locals = locals.sort();
         el.title = locals.join('\n');
-        el.onclick = function () {
+        el.onclick = function() {
             alert('Local variables: ' + locals.join(', '));
         };
     }
@@ -137,10 +292,11 @@ window.onload = function() {
         initLocalsButton('localizer', ['module', 'exports']);
     }
 
-    function initObjectEditor(type) {
-        var scriptList = scripts[type],
+    function initObjectEditor(type, data) {
+        var typedScripts = scripts[type] = data.split(/\s*\/\* ✂ \*\/\s*/),
             dropdownEl = document.getElementById(type + '-dropdown'),
-            resetEl = dropdownEl.nextElementSibling,
+            resetEl = document.getElementById('reset-' + type),
+            deleteEl = document.getElementById('delete-' + type),
             scriptEl = document.getElementById(type + '-script'),
             addButtonEl = dropdownEl.parentElement.querySelector('.api'),
             prefix = type + '_',
@@ -148,13 +304,13 @@ window.onload = function() {
             newScript;
 
         // STEP 1: Save default scripts to local storage not previously saved
-        scriptList.map(extractName).sort().forEach(function (key) {
-            var script = scriptList.find(isScriptByName.bind(null, key));
-            if (key === NEW) {
-                dropdownEl.add(new Option(key));
+        typedScripts.map(extractName).sort().forEach(function(name) {
+            var script = getDefaultScript(type, name);
+            if (name === NEW) {
+                dropdownEl.add(new Option(name));
                 newScript = scriptEl.value = script;
-            } else if (!localStorage.getItem(prefix + key)) {
-                localStorage.setItem(prefix + key, script);
+            } else if (!localStorage.getItem(prefix + name)) {
+                localStorage.setItem(prefix + name, script);
             }
         });
 
@@ -168,30 +324,58 @@ window.onload = function() {
 
         // STEP 3: Reset drop-down to first item: "(New)"
         dropdownEl.selectedIndex = 0;
-        resetEl.style.display = 'none';
+        enableResetAndDeleteIcons(type, NEW);
 
-        resetEl.onclick = resetObjectEditor;
+        // STEP 4: Assign handlers
+        resetEl.onclick = resetObject;
+        deleteEl.onclick = deleteObject;
 
-        dropdownEl.onchange = function () {
-            var name = this.value;
+        var name = NEW;
+        dropdownEl.onchange = function() {
+            var newName = this.value;
+            var savedScript = localStorage.getItem(prefix + name) || getDefaultScript(type, name);
+            var editedScript = document.getElementById(type + '-script').value;
 
-            if (name === NEW) {
-                scriptEl.value = newScript;
-                resettable(type);
-            } else {
-                scriptEl.value = localStorage.getItem(prefix + name);
-                resettable(type, name);
+            if (!savedScript || savedScript === editedScript || confirm('Discard unsaved changes?')) {
+                name = newName;
+                if (name === NEW) {
+                    scriptEl.value = newScript;
+                    enableResetAndDeleteIcons(type);
+                } else {
+                    scriptEl.value = localStorage.getItem(prefix + name);
+                    enableResetAndDeleteIcons(type, name);
+                }
+            } else if (savedScript) {
+                this.value = name;
             }
         };
 
-        addButtonEl.onclick = function () {
+        scriptEl.oninput = function() {
+            enableResetAndDeleteIcons(type, dropdownEl.value);
+        };
+
+        addButtonEl.onclick = function() {
             save(scriptEl.value, dropdownEl);
+            grid.repaint();
         };
     }
 
-    function resettable(type, name) {
-        var hasDefaultScript = name && !!scripts[type].find(isScriptByName.bind(null, name));
-        document.getElementById('reset-' + type).style.display = hasDefaultScript ? null : 'none'; // null reveals start value
+    function enableResetAndDeleteIcons(type, name) {
+        var resetEl = document.getElementById('reset-' + type);
+        var deleteEl = document.getElementById('delete-' + type);
+
+        if (name) {
+            var defaultScript = name && getDefaultScript(type, name);
+            var editedScript = document.getElementById(type + '-script').value;
+            var alteredFromDefault = defaultScript && defaultScript !== editedScript;
+
+            resetEl.classList.toggle('disabled', !alteredFromDefault);
+            deleteEl.classList.toggle('disabled', !name || defaultScript);
+        }
+    }
+
+    function getDefaultScript(type, name) {
+        return scripts[type].find(isScriptByName.bind(null, name));
     }
 
     function isScriptByName(name, script) {
@@ -203,36 +387,15 @@ window.onload = function() {
         return match[1] || match[2] || match[3] || match[4];
     }
 
-    function putJSON(key, json) {
-        document.getElementById(key).value = JSON.stringify(json, undefined, 2)
+    function stringifyAndUnquoteKeys(json) {
+        return JSON.stringify(json, undefined, 2)
             .replace(/(  +)"([a-zA-Z$_]+)"(: )/g, '$1$2$3'); // un-quote keys
-    }
-
-    function callApi(methodName, id, confirmation) {
-        var jsonEl = document.getElementById(id), value = jsonEl.value;
-
-        localStorage.setItem(id, value);
-
-        // L-value must be inside eval because R-value beginning with '{' is eval'd as BEGIN block
-        // used eval because JSON.parse rejects unquoted keys
-        eval('value =' + value);
-
-        var params = [value];
-        if (methodName === 'setData') {
-            params.push({ schema: [] });
-        }
-
-        grid[methodName].apply(grid, params);
-
-        if (confirmation) {
-            feedback(jsonEl.parentElement, confirmation);
-        }
     }
 
     function saveCellEditor(script, select) {
         var cellEditors = grid.cellEditors;
         var editorNames = Object.keys(cellEditors.items);
-        var editors = editorNames.map(function (name) {
+        var editors = editorNames.map(function(name) {
             return cellEditors.items[name];
         });
         var exports = {}, module = { exports: exports };
@@ -266,7 +429,7 @@ window.onload = function() {
 
         addOptionInAlphaOrder(select, name);
 
-        resettable('editor', name);
+        enableResetAndDeleteIcons('editor', name);
 
         if (select) {
             feedback(select.parentElement);
@@ -297,7 +460,7 @@ window.onload = function() {
 
         addOptionInAlphaOrder(select, name);
 
-        resettable('localizer', name);
+        enableResetAndDeleteIcons('localizer', name);
 
         if (select) {
             feedback(select.parentElement);
@@ -311,7 +474,7 @@ window.onload = function() {
 
         var optionEls = Array.prototype.slice.call(el.options);
 
-        var index = optionEls.findIndex(function (optionEl) {
+        var index = optionEls.findIndex(function(optionEl) {
             return optionEl.textContent === text;
         });
         if (index >= 0) {
@@ -319,7 +482,7 @@ window.onload = function() {
             return; // already in dropdown
         }
 
-        var firstOptionGreaterThan = optionEls.find(function (optionEl) {
+        var firstOptionGreaterThan = optionEls.find(function(optionEl) {
             return optionEl.textContent > text;
         });
 
@@ -337,9 +500,10 @@ window.onload = function() {
         }
         el.innerText = confirmation;
         el.style.display = 'inline';
-        setTimeout(function () {
+        setTimeout(function() {
             el.style.display = 'none';
         }, 750 + 50 * confirmation.length);
     }
 
+    window.callApi = callApi; // for access from index.html `onclick` handlers
 };
